+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `Select` - removed value from internal state
 - `Autocomplete` - removed defaultValue from internal state
 
+### Changed
+- `Autocomplete` Add disabled option to items array
+
 
 ## v6.0.6 - 2021-10-08